macOS: move SearchState into SurfaceView+SearchState.swift
Ghostty.SurfaceView.swift is 2714 lines after absorbing OSSurfaceView. Move
the self-contained extensions out into the SurfaceView+Feature.swift
convention already used by SurfaceView+Image.swift and
SurfaceView+Transferable.swift, starting with the SearchState model and the
two search navigation helpers.

This also pairs the file with the existing SurfaceView+SearchStateTests.swift.

Pure code motion; the only change is dropping the MARK header the file name
now carries.
